code/FixMatch-pytorch/my_requirements.txt
#absl-py==0.10.0
#alabaster==0.7.12
#argh==0.26.2
#asn1crypto==1.3.0
#astroid==2.4.2
#astropy==4.0.1.post1
#atomicwrites==1.3.0
#attrs==19.3.0
#autopep8==1.4.4
#autovizwidget==0.15.0
#Babel==2.8.0
#backcall==0.1.0
#backports.functools-lru-cache==1.6.1
#backports.shutil-get-terminal-size==1.0.0
#backports.tempfile==1.0
#backports.weakref==1.0.post1
#beautifulsoup4==4.8.2
#bitarray==1.2.1
#bkcharts==0.2
#bleach==3.1.4
#bokeh==2.0.1
#boto==2.49.0
#boto3==1.14.45
#botocore==1.17.45
#Bottleneck==1.3.2
#cachetools==4.1.1
#certifi==2020.6.20
#cffi==1.14.0
#chardet==3.0.4
#click==7.1.1
#cloudpickle==1.3.0
#colorama==0.4.3
#contextlib2==0.6.0.post1
#cryptography==2.8
#cycler==0.10.0
Cython==0.29.15
#cytoolz==0.10.1
#dask==2.14.0
#decorator==4.4.2
#defusedxml==0.6.0
#diff-match-patch==20181111
#distributed==2.14.0
#docutils==0.15.2
#entrypoints==0.3
#environment-kernels==1.1.1
#et-xmlfile==1.0.1
#fastcache==1.1.0
#filelock==3.0.12
#flake8==3.7.9
#Flask==1.1.1
#fsspec==0.7.1
#future==0.18.2
#gevent==1.4.0
#glob2==0.7
#gmpy2==2.0.8
#google-auth==1.22.1
#google-auth-oauthlib==0.4.1
#greenlet==0.4.15
#grpcio==1.33.1
#h5py==2.10.0
#hdijupyterutils==0.15.0
#HeapDict==1.0.1
#html5lib==1.0.1
#hypothesis==5.8.3
#idna==2.9
#imageio==2.8.0
#imagesize==1.2.0
#importlib-metadata==1.5.0
#intervaltree==3.0.2
#ipykernel==5.1.4
#ipyparallel==6.2.4
#ipython==7.13.0
#ipython-genutils==0.2.0
#ipywidgets==7.5.1
#isort==4.3.21
#itsdangerous==1.1.0
#jdcal==1.4.1
#jedi==0.15.2
#jeepney==0.4.3
#Jinja2==2.11.1
#jmespath==0.9.4
#joblib==0.14.1
json5==0.9.4
jsonlines==1.2.0
jsonschema==3.2.0
keyring==21.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.3.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
openpyxl==3.0.3
packaging==20.3
pandas
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.9.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.13.0
protobuf3-to-dict==0.1.5
psutil==5.7.0
psycopg2==2.7.5
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pydocstyle==4.0.1
pyflakes==2.1.1
pygal==2.4.0
Pygments==2.6.1
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.1
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
pytorch-pretrained-bert==0.6.2
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.7.0
qtconsole==4.7.2
QtPy==1.9.0
regex==2020.10.15
requests==2.23.0
requests-kerberos==0.12.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.16.0
rsa==4.6
ruamel-yaml==0.15.87
s3fs==0.4.0
s3transfer==0.3.3
sagemaker==1.72.0
scikit-image==0.16.2
scikit-learn
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
smdebug-rulesconfig==0.1.4
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0
sparkmagic==0.15.0
Sphinx
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
spyder==4.1.2
spyder-kernels==1.9.0
SQLAlchemy==1.3.15
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl
toml==0.10.1
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.44.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
zict==2.0.0
zipp==2.2.0
